Implement Peak Flow Meter
**Not Complete**

Peakflow meter will connect to device and create a txt file of the data
from the device. Currently working on fail safe checks to make it easier
on the user to know when the device is disconnected. all work but the
automatic disconnect when the device powers off.

Yet to do -

Pull date and set that to text view on page of time that last data
upload(file addition) occured.

add a instructional pop up in the settings menu instead of following the
same as all other pages. So that the user can see a walk through of how
to use the peakflow meter (Possible PDF creation and display? or image)