Bluetooth Exceptions
Fixed so program does not crash but pushes you to bluetooth settings in
situation where bluetooth is disabled or bluetooth devices are not
connected. Working on a popup window for instructions for the user  from
the options menu. Still have buttons available for testing purposes but
will be removed when attempt connection/collect data is functional.